+++ relay-config/src/lib.rs
@@ -6,10 +6,12 @@
 )]
 #![allow(clippy::derive_partial_eq_without_eq)]
 
+pub mod aggregator;
 mod byte_size;
 mod config;
 mod upstream;
 
+pub use crate::aggregator::{AggregatorServiceConfig, ScopedAggregatorConfig};
 pub use crate::byte_size::*;
 pub use crate::config::*;
 pub use crate::upstream::*;
